// frontend/src/app/core/api/models/courseDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CourseDto { 
    id?: string;
    teacherId?: string;
    courseName?: string;
    syllabus?: string;
    credits?: number;
    description?: string;
}